Enabled service call to start and stop dmp teach trajectory record
