+++ costar_robot/costar_dmp/scripts/dmp_server.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+
+import rospy
+from costar_dmp import CostarDMP
+
+rospy.init_node('costar_dmp_server')
+server = CostarDMP()
+
+rate = rospy.Rate(10)
+
+while not rospy.is_shutdown():
+
+	server.dmp_tick()
+	rate.sleep()
