+++ costar_robot/costar_dmp/src/costar_dmp/__init__.py
@@ -1,0 +1,2 @@
+import commands
+from dmp_server import CostarDMP
