+++ costar_robot/costar_dmp/srv/StartRec.srv
@@ -1,0 +1,5 @@
+bool collecting
+
+---
+
+bool status
